Utterance start and end times are now float instead of double in order to save some memory used due to pointer size diff

+++ analysis/src/main/java/se/kth/speech/coin/tangrams/analysis/PlayerGameContextFactory.java
@@ -27,7 +27,7 @@ final class PlayerGameContextFactory {
 		this.playerGameHistoryGetter = playerGameHistoryGetter;
 	}
 
-	public Stream<GameContext> create(final double startTime, final double endTime,
+	public Stream<GameContext> create(final float startTime, final float endTime,
 			final String perspectivePlayerId) {
 		UtteranceSelectedEntityDescriptionWriter.LOGGER.debug("Creating a context based on the logged game history from the perspective of player \"{}\".",
 				perspectivePlayerId);
